added cython compiled batch version of bounding box computation
